+++ src/pages/account/index.tsx
@@ -7,9 +7,9 @@ import { AccountOverview } from "@src/components/display/fewlines/AccountOvervie
 import { Container } from "@src/components/display/fewlines/Container";
 import { H1 } from "@src/components/display/fewlines/H1/H1";
 import { oauth2Client, config } from "@src/config";
-import { deviceBreakpoints } from "@src/design-system/theme";
-import { withSSRLogger } from "@src/middleware/withSSRLogger";
-import withSession from "@src/middleware/withSession";
+import { deviceBreakpoints } from "@src/design-system/theme/lightTheme";
+import { withSSRLogger } from "@src/middlewares/withSSRLogger";
+import withSession from "@src/middlewares/withSession";
 import { getUser } from "@src/utils/getUser";
 import { refreshTokens } from "@src/utils/refreshTokens";
 import Sentry, { addRequestScopeToSentry } from "@src/utils/sentry";
