+++ src/pages/api/auth-connect/verify-validation-code.ts
@@ -7,9 +7,9 @@ import type { ExtendedRequest } from "@src/@types/ExtendedRequest";
 import { HttpVerbs } from "@src/@types/HttpVerbs";
 import { config } from "@src/config";
 import { MongoNoDataReturned, TemporaryIdentityExpired } from "@src/errors";
-import { withAPIPageLogger } from "@src/middleware/withAPIPageLogger";
-import { withMongoDB } from "@src/middleware/withMongoDB";
-import withSession from "@src/middleware/withSession";
+import { withAPIPageLogger } from "@src/middlewares/withAPIPageLogger";
+import { withMongoDB } from "@src/middlewares/withMongoDB";
+import withSession from "@src/middlewares/withSession";
 import { fetchJson } from "@src/utils/fetchJson";
 import Sentry, { addRequestScopeToSentry } from "@src/utils/sentry";
