+++ src/pages/api/oauth/callback.ts
@@ -4,11 +4,11 @@ import { Handler } from "next-iron-session";
 
 import type { ExtendedRequest } from "@src/@types/ExtendedRequest";
 import type { AccessToken } from "@src/@types/oauth2/OAuth2Tokens";
-import { findOrInsertUser } from "@src/command/findOrInsertUser";
+import { findOrInsertUser } from "@src/commands/findOrInsertUser";
 import { oauth2Client, config } from "@src/config";
-import { withAPIPageLogger } from "@src/middleware/withAPIPageLogger";
-import { withMongoDB } from "@src/middleware/withMongoDB";
-import withSession from "@src/middleware/withSession";
+import { withAPIPageLogger } from "@src/middlewares/withAPIPageLogger";
+import { withMongoDB } from "@src/middlewares/withMongoDB";
+import withSession from "@src/middlewares/withSession";
 import Sentry, { addRequestScopeToSentry } from "@src/utils/sentry";
 
 const handler: Handler = async (
