+++ src/pages/api/oauth/refresh-tokens.ts
@@ -6,9 +6,9 @@ import type { Handler } from "next-iron-session";
 import { refreshTokensFlow } from "@lib/commands/refreshTokensFlow";
 import type { ExtendedRequest } from "@src/@types/ExtendedRequest";
 import { MongoUpdateError } from "@src/errors";
-import { withAPIPageLogger } from "@src/middleware/withAPIPageLogger";
-import { withMongoDB } from "@src/middleware/withMongoDB";
-import withSession from "@src/middleware/withSession";
+import { withAPIPageLogger } from "@src/middlewares/withAPIPageLogger";
+import { withMongoDB } from "@src/middlewares/withMongoDB";
+import withSession from "@src/middlewares/withSession";
 import Sentry, { addRequestScopeToSentry } from "@src/utils/sentry";
 
 const handler: Handler = async (
